Various small fixes and improvements (#228)
* Change label

* Adjust example 37

* Revert "Adjust example 37"

This reverts commit 1c99274975e6f541d1571f17ad590b9cca509957.

* Adjust example

* Fix null pointer exception when configuring custom generalization
schemes to use with differential privacy

* Fix example

* Fix messages

* Do not allow the specification of the heuristic step limit
Integer.MAX_VALUE via the GUI

* Improve rendering of information loss based on differential privacy
scores

* Fix

* Specify search budget as percentage of the overall privacy parameter
epsilon when using the GUI

* Restore bugfix

* Add comments

* Simplify calculation of the absolute search budget

* Reset range of delta

* Adjust ranges of parameters for differential privacy

* Refactoring and small improvements

+++ src/main/org/deidentifier/arx/metric/v2/ILScore.java
@@ -98,7 +98,7 @@ public class ILScore extends InformationLoss<BigFraction> {
 
     @Override
     public String toString() {
-        return String.valueOf(this.value);
+        return String.valueOf(this.value.doubleValue());
     }
 
     /**
